Added a filter to remove only discussion comments and keep original channel messages that match antispam

+++ bot.py
@@ -1,5 +1,7 @@
 import logging
+import operator
 import os
+from functools import reduce
 from telegram import Message, Update
 from telegram.ext import CallbackContext, MessageHandler, Updater
 from telegram.ext.filters import Filters, MessageFilter
@@ -12,6 +14,8 @@ def get_message(update: Update) -> Message:
 
 def delete(update: Update, context: CallbackContext):
     message = get_message(update)
+    if message is None:
+        return
 
     message.bot.delete_message(
         message_id=message.message_id,
@@ -32,6 +36,25 @@ class ContainsLink(MessageFilter):
         return len(self.extractor.find_urls(message.text)) >= 1
 
 
+class ChatMessageOnly(MessageFilter):
+    def filter(self, message: Message) -> bool:
+        return message.forward_from_message_id is None
+
+
+def with_default_filters(*filters):
+    """Apply default filters to the given filter classes"""
+    default_filters = [
+        Filters.text,
+        ChatMessageOnly(),
+    ]
+    return reduce(operator.and_, [*default_filters, *filters])  # МАМА Я УМЕЮ ФУНКЦИОНАЛЬНО ПРОГРАММИРОВАТЬ
+
+
+def delete_messages_that_match(*filters) -> MessageHandler:
+    """Sugar for quick adding delete message callbacks"""
+    return MessageHandler(callback=delete, filters=with_default_filters(*filters))
+
+
 def in_heroku() -> bool:
     return os.getenv('HEROKU_APP_NAME', None) is not None
 
@@ -48,12 +71,17 @@ def main():
     app_name = os.getenv('HEROKU_APP_NAME')
 
     bot = Updater(token=bot_token)
-    bot.dispatcher.add_handler(MessageHandler(callback=delete, filters=Filters.text & ContainsTelegramContact()))
-    bot.dispatcher.add_handler(MessageHandler(callback=delete, filters=Filters.text & ContainsLink()))
+    bot.dispatcher.add_handler(
+        delete_messages_that_match(ContainsTelegramContact()),
+    )
+    bot.dispatcher.add_handler(
+        delete_messages_that_match(ContainsLink()),
+    )
 
     if not in_heroku():
         enable_logging()
         bot.start_polling()
+
     else:
         bot.start_webhook(
             listen='0.0.0.0',
